gpu: nvgpu: pmu: fix MISRA 10.3 issues in pmu_perfmon
MISRA Rule 10.3 prohibits implicit assignment of objects of narrower
size or essential type. This fixes MISRA 10.3 violations in pmu_ipc.c

JIRA NVGPU-2841

Reviewed-on: https://git-master.nvidia.com/r/2027769
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/pmu.h
@@ -394,7 +394,7 @@ struct nvgpu_pmu {
 	union {
 		struct pmu_perfmon_counter_v2 perfmon_counter_v2;
 	};
-	u32 perfmon_state_id[PMU_DOMAIN_GROUP_NUM];
+	u8 perfmon_state_id[PMU_DOMAIN_GROUP_NUM];
 
 	bool initialized;
